backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
requests==2.31.0
aiofiles==23.2.1
firebase-admin==6.2.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
email-validator==2.1.0
jinja2==3.1.2